requirements.txt
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.13.0
PyYAML==5.3.1
